refactor: merge state into solving
