+++ expertsystem/__init__.py
@@ -19,10 +19,9 @@ __all__ = [
     "amplitude",
     "data",
     "io",
-    "state",
     "solving",
     "ui",
 ]
 
 
-from . import amplitude, data, io, solving, state, ui
+from . import amplitude, data, io, solving, ui
