+++ expertsystem/solving/combinatorics.py
