+++ expertsystem/solving/properties.py
