--- expertsystem/state/__init__.py
@@ -1,8 +1,0 @@
-"""All modules related to a quantum mechanical state."""
-
-__all__ = [
-    "combinatorics",
-    "properties",
-]
-
-from . import combinatorics, properties
