--- tests/unit/state/__init__.py
@@ -1,1 +1,0 @@
-"""Required to ignore mypy errors in the :file:`tests` folder."""
